refactor(binary-codec): updated testing coverage on definitions package with unnecessary code removal
