+++ binary-codec/definitions/field_instance_test.go
@@ -1,0 +1,19 @@
+package definitions
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestDefinitions_CreateFieldHeader(t *testing.T) {
+	loadDefinitions()
+	require.Equal(t, FieldHeader{
+		TypeCode:  1,
+		FieldCode: 2,
+	}, Get().CreateFieldHeader(1, 2))
+}
+
+func TestFieldInstanceMap_CodecDecodeSelf(t *testing.T) {
+	loadDefinitions()
+}
