+++ binary-codec/definitions/init.go
@@ -1,8 +1,5 @@
 package definitions
 
 func init() {
-	err := loadDefinitions()
-	if err != nil {
-		return
-	}
+	loadDefinitions()
 }
